Refactor graphs: create separate CrimesByGender and CrimesByGunStolen components.
