+++ server/src/routes/Gun.ts
@@ -22,4 +22,26 @@ router.get('/types', async (req: Request, res: Response) => {
   }
 });
 
+/**
+ * Number of incidents caused by stolen/not stolen guns.
+ */
+router.get(
+  '/numberOfCrimesByGunStolen',
+  async (req: Request, res: Response) => {
+    try {
+      const stolen = await query(
+        `SELECT stolen, COUNT(id) AS n_crimes
+        FROM ${Gun}
+        GROUP BY stolen`
+      );
+      return res.status(OK).json(stolen);
+    } catch (err) {
+      logger.error(err.message, err);
+      return res.status(BAD_REQUEST).json({
+        error: err.message,
+      });
+    }
+  }
+);
+
 export default router;
